Lower cpu and memory, change memory to memoryReservation
Since bursting goes past these limits as long as there are free resources on the container, it seems fine to set these lower for now.

Memory reservation is the soft limit parameter, meaning that the container won't be killed if it goes over this limit. Plain memory was the hard parameter.